增加了对IC卡充值的支持 增加了充值记录查询API

+++ src/server/model/soda/account.go
@@ -4,7 +4,7 @@ import "maizuo.com/soda-manager/src/server/model"
 
 type Account struct {
 	model.Model
-	UserId int64  `json:"userId,omitempty"`
+	UserId int  `json:"userId,omitempty"`
 	App    string `json:"app,omitempty"`
 	Key    string `json:"key,omitempty"`
 	Extra  string `json:"extra,omitempty"`
